% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SignalBind.R
\name{bind_signal_data}
\alias{bind_signal_data}
\title{Signal Clean Up Draft
A function to bind and process Logie signal data
This function allows you to bind together mulitple signal data files, remove errors and produce a master datafile.}
\usage{
bind_signal_data(path.to.folder, site, year)
}
\arguments{
\item{path.to.folder}{This is the file path for the folder that contains all data files for processing.}

\item{year}{Year of counter operation.}

\item{site.name}{Name of the study river.}
}
\description{
Signal Clean Up Draft
A function to bind and process Logie signal data
This function allows you to bind together mulitple signal data files, remove errors and produce a master datafile.
}
\keyword{Logie}

