% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CounterBind.R
\name{counter.data.cleanup}
\alias{counter.data.cleanup}
\title{Counter Clean Up
A function to bind and process Logie counter data
This function allows you to bind together mulitple counter data files, remove errors and produce a master datafile.}
\usage{
counter.data.cleanup(path.to.folder, no.channels, site.name, year)
}
\arguments{
\item{path.to.folder}{This is the file path for the folder that contains all data files for processing.}

\item{no.channels}{This is the number of counter channels that were operated.}

\item{site.name}{Name of the study river.}

\item{year}{Year of counter operation.}
}
\description{
Counter Clean Up
A function to bind and process Logie counter data
This function allows you to bind together mulitple counter data files, remove errors and produce a master datafile.
}

