% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EventsHourChPlot.R
\name{events.plot}
\alias{events.plot}
\title{A function that plots events per hour of Logie counter data}
\usage{
events.plot(dataset, first.day, site, year)
}
\arguments{
\item{first.day}{This is the first day of the dataset you want to use. This parameter needs to be specified in julian day format.}

\item{site}{Name of the study river.}

\item{year}{Year of counter operation.}

\item{data}{This is the dataset used to create the plots.}
}
\description{
This function plots the number of events per hour channel and day for Logie counter data by channel
}
\keyword{Events}

