% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HistPlotAll.R
\name{hist_all_records}
\alias{hist_all_records}
\title{A function that plots histograms of Logie counter data}
\usage{
hist_all_records(dataset, first_day = NULL, last_day = NULL)
}
\arguments{
\item{dataset}{This is the dataset used to create the histograms.}

\item{first_day}{This is the first day of the dataset you want to use. This parameter needs to be specified in julian day format. Defaults to the first day in the dataset}

\item{last_day}{This is the last day of the dataset you want to use. This parameter needs to be specified in julian day format. Defaults to the last day in the dataset.}
}
\value{
Generates histograms of peak signal size for up counts, down counts and events for each counter channel.
}
\description{
This function plots historgrams of up, down and event counts for Logie counter data by channel
}
\keyword{Histogram}

