% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PSSHourPlot.R, R/PSSPlot.R
\name{pss.date.plot}
\alias{pss.date.plot}
\title{A function that plots the peak signal size (PSS) by hour of Logie counter data}
\usage{
pss.date.plot(dataset, first.day, site, year, low.thresh, up.thresh)

pss.date.plot(dataset, first.day, site, year, low.thresh, up.thresh)
}
\arguments{
\item{dataset}{This is the dataset used to create the plots.}

\item{first.day}{This is the first day of the dataset you want to use. This parameter needs to be specified in julian day format.}

\item{site}{Name of the study river.}

\item{year}{Year of counter operation.}

\item{low.thresh}{is the counters lower threshold PSS value.}

\item{up.thresh}{is the counters upper threshold PSS value.}

\item{dataset}{This is the dataset used to create the plots.}

\item{first.day}{This is the first day of the dataset you want to use. This parameter needs to be specified in julian day format.}

\item{site}{Name of the study river.}

\item{year}{Year of counter operation.}

\item{low.thresh}{is the counters lower threshold PSS value.}

\item{up.thresh}{is the counters upper threshold PSS value.}
}
\description{
This function plots the PSS by hour for Logie counter data

This function plots the PSS by day and time for Logie counter data
}
\keyword{Events}

