% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HistPlot.R
\name{record.hist}
\alias{record.hist}
\title{A function that plots histograms of Logie counter data}
\usage{
record.hist(dataset, first.day, site, year)
}
\arguments{
\item{dataset}{This is the dataset used to create the histograms.}

\item{first.day}{This is the first day of the dataset you want to use. This parameter needs to be specified in julian day format.}

\item{site}{Name of the study river.}

\item{year}{Year of counter operation.}
}
\description{
This function plots historgrams of up, down and event counts for Logie counter data by channel
}
\keyword{Histogram}

