% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PSSHourPlot.R
\docType{data}
\name{sunrise}
\alias{sunrise}
\title{A function that plots the peak signal size (PSS) by hour of Logie counter data}
\format{\preformatted{ num 8
}}
\usage{
sunrise
}
\arguments{
\item{dataset}{This is the dataset used to create the plots.}

\item{first.day}{This is the first day of the dataset you want to use. This parameter needs to be specified in julian day format.}

\item{site}{Name of the study river.}

\item{year}{Year of counter operation.}

\item{low.thresh}{is the counters lower threshold PSS value.}

\item{up.thresh}{is the counters upper threshold PSS value.}

\item{sunrise}{is the aproximate time of sunrise needs to be an integer value between 0 and 12}

\item{sunset}{is the aproximate time of sunrise needs to be an integer value 12 and 24}
}
\description{
This function plots the PSS by hour for Logie counter data
}
\keyword{Events}

