% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CounterBind.R
\name{bind_counter_data}
\alias{bind_counter_data}
\title{bind_counter_data
bind individual data files, clean up data, and return a master data file.
A function to bind and process Logie counter data
This function allows you to bind together mulitple counter data files, remove errors and produce a master datafile.}
\usage{
bind_counter_data(path_to_folder = ".", no_channels, site, year, max_signal,
  rows_rm = FALSE)
}
\arguments{
\item{path_to_folder}{This is the file path for the folder that contains all data files for processing.}

\item{no_channels}{This is the number of counter channels that were operated.}

\item{site}{Name of the study river.}

\item{year}{Year of counter operation.}

\item{max_signal}{The maximum signal size.}
}
\description{
bind_counter_data
bind individual data files, clean up data, and return a master data file.
A function to bind and process Logie counter data
This function allows you to bind together mulitple counter data files, remove errors and produce a master datafile.
}

