% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HistPlot.R
\name{hist_records}
\alias{hist_records}
\title{A function that plots histograms of Logie counter data for the specified direction}
\usage{
hist_records(dataset, direction, day_one = NULL)
}
\arguments{
\item{dataset}{This is the dataset used to create the histograms.}

\item{direction}{This is the type of record ("U", "D", or "E") to be plotted.}

\item{day_one}{This is the first day of the dataset you want to use, defaults to the first day found in the dataset. This parameter needs to be specified in julian day format.}
}
\value{
Generates a histograms of peak signal size for either up counts, down counts and events for each counter channel.
}
\description{
This function plots historgrams of up, down and event counts for Logie counter data by channel as specified by the user.
}
\keyword{Histogram}

