% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PSSHourPlot.R
\name{plot_pss_hour}
\alias{plot_pss_hour}
\title{A function that plots the peak signal size (PSS) by hour of Logie counter data}
\usage{
plot_pss_hour(dataset, first_day = NULL, last_day = NULL,
  low_thresh = NULL, up_thresh = NULL, ch = NULL)
}
\arguments{
\item{dataset}{This is the dataset used to create the plots.}

\item{first_day}{This is the first day of the dataset you want to use. This parameter needs to be specified in julian day format. Defaults to the first day in the dataset}

\item{last_day}{This is the last day of the dataset you want to use. This parameter needs to be specified in julian day format. Defaults to the last day in the dataset.}

\item{low_thresh}{is the counters lower threshold PSS value. Defaults to 0.}

\item{up_thresh}{is the counters upper threshold PSS value. Defaults to 130.}

\item{ch}{is the channel to be plotted. Defaults to all channels.}
}
\description{
This function plots the PSS by hour for Logie counter data
}
\keyword{Events}

